# tests/snapshots/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "ApplyRequest": {
        "properties": {
          "cost_estimate": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Estimate"
          },
          "job_name": {
            "title": "Job Name",
            "type": "string"
          },
          "preflight_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preflight Hash"
          },
          "workspace_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Dir"
          }
        },
        "required": [
          "job_name"
        ],
        "title": "ApplyRequest",
        "type": "object"
      },
...
      "CancelRequest": {
        "properties": {
          "job_name": {
            "title": "Job Name",
            "type": "string"
          },
          "workspace_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Dir"
          }
        },
        "required": [
          "job_name"
        ],
        "title": "CancelRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunRequest": {
        "properties": {
          "backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "local",
            "title": "Backend"
          },
          "job_name": {
            "title": "Job Name",
            "type": "string"
          },
          "workspace_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Dir"
          }
        },
        "required": [
          "job_name"
        ],
        "title": "RunRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Data Platform Copilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/build/": {
      "post": {
        "operationId": "build_pipeline_api_v1_build__post",
        "parameters": [
          {
            "in": "query",
            "name": "base_ref",
            "required": false,
            "schema": {
              "default": "upstream/main",
              "title": "Base Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_intelligence",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Intelligence",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Pipeline",
        "tags": [
          "Build"
        ]
      }
    },
...
    "/api/v2/build/": {
      "post": {
        "operationId": "build_pipeline_api_v2_build__post",
        "parameters": [
          {
            "in": "query",
            "name": "base_ref",
            "required": false,
            "schema": {
              "default": "upstream/main",
              "title": "Base Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_intelligence",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Intelligence",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Pipeline",
        "tags": [
          "Build"
        ]
      }
    },
    "/api/v2/build/apply": {
      "post": {
        "operationId": "apply_build_api_v2_build_apply_post",
        "parameters": [
          {
            "in": "query",
            "name": "contract_hash",
            "required": true,
            "schema": {
              "title": "Contract Hash",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Apply Build Api V2 Build Apply Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Apply Build",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/v2/build/cancel": {
      "post": {
        "operationId": "cancel_build_api_v2_build_cancel_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Tenant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cancel Build Api V2 Build Cancel Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Build",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/v2/build/history/{job_name}": {
      "get": {
        "operationId": "history_api_v2_build_history__job_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_dir",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Dir"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response History Api V2 Build History  Job Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/v2/build/run": {
      "post": {
        "operationId": "run_build_api_v2_build_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "contract_hash",
            "required": true,
            "schema": {
              "title": "Contract Hash",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant"
            }
          },
          {
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Request-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Run Build Api V2 Build Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Build",
        "tags": [
          "execution"
        ]
      }
    },
    "/api/v2/build/status/{job_name}": {
      "get": {
        "operationId": "status_api_v2_build_status__job_name__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "workspace_dir",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Dir"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Status Api V2 Build Status  Job Name  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Status",
        "tags": [
          "execution"
        ]
      }
    },
...
    "/api/v3/build/run": {
      "post": {
        "operationId": "run_build_api_v3_build_run_post",
        "parameters": [
          {
            "in": "query",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "contract_hash",
            "required": true,
            "schema": {
              "title": "Contract Hash",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plan_hash",
            "required": true,
            "schema": {
              "title": "Plan Hash",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "intelligence_hash",
            "required": true,
            "schema": {
              "title": "Intelligence Hash",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "policy_eval_hash",
            "required": true,
            "schema": {
              "title": "Policy Eval Hash",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Build",
        "tags": [
          "build_v3"
        ]
      }
    },
...
    "/build/": {
      "post": {
        "operationId": "build_pipeline_build__post",
        "parameters": [
          {
            "in": "query",
            "name": "base_ref",
            "required": false,
            "schema": {
              "default": "upstream/main",
              "title": "Base Ref",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "include_intelligence",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Intelligence",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Pipeline",
        "tags": [
          "Build"
        ]
      }
    },
...
  }
}